+++ p2pserver/message/inventory.go
